web.template.yml: remove copy-env script.
This script needlessly writes out DISCOURSE_* environment variables to
disk, which is not required by the Discourse application because it will
already be able to read the secrets from the environment. Having
secrets stored on disk when they are not required to be provides an attack
vector in the event of remote code execution within the app container.