Strip out env variables when committing.
By default, docker stores all of the environment variables present in a
container in the resulting image when running a `commit`. This can be
useful, however means that any sensitive environment variables are
visible via `docker inspect`. We always provide the environment
to the container at runtime, so we don't need to store the variable at
all at rest.

This change causes any environment variables with the prefix
`DISCOURSE_` to be unset when committed.